refactor: move file parsing logic out of __init__.py
